moved component styling into seperate files
