+++ hour-tracker/src/components/Sider/CustomSider.css
@@ -1,0 +1,28 @@
+
+  .sider-content {
+    padding: 0.5em 0.5em;
+    margin: 5em 0.5em;
+    display: flex;
+    justify-content: center;
+    height: 90%;
+  }
+  
+  .sider-profile {
+    margin-top: 3em;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .sider-profile-details {
+    margin-top: 1.5em;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+  
+  .user-name {
+    font-size: 20px;
+    margin-bottom: -0.8em;
+  }
+
